# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Car Configurations API",
        "description": "API for managing cars, configurations, options and prices",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/prices": {
            "get": {
                "tags": [
                    "Prices"
                ],
                "summary": "Get all prices",
                "operationId": "2e9eb9dd53b4217762ae89f858505c78",
                "responses": {
                    "200": {
                        "description": "List of all prices",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {}
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Prices"
                ],
                "summary": "Create a new price",
                "operationId": "2a4eff300e8d57617fa7b5ae324048dc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "configuration_id",
                                    "price",
                                    "start_date"
                                ],
                                "properties": {
                                    "configuration_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 35000
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-01-01"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-12-31"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Price created successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/prices/{id}": {
            "get": {
                "tags": [
                    "Prices"
                ],
                "summary": "Get a single price",
                "operationId": "1a2f37f16e6736fe50f4d1813aa798d1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the price to get",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Price details",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Price not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Prices"
                ],
                "summary": "Update a price",
                "operationId": "cb8d0d4c496a8d7c7dca39fb7abe4596",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the price to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "configuration_id",
                                    "price",
                                    "start_date"
                                ],
                                "properties": {
                                    "configuration_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 35000
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-01-01"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-12-31"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Price updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Price not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Prices"
                ],
                "summary": "Delete a price",
                "operationId": "7d4527622692561d05e145994af53a94",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the price to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Price deleted successfully"
                    },
                    "404": {
                        "description": "Price not found"
                    }
                }
            }
        }
    },
...
}